requirements.txt
Django==2.2.1
djangorestframework==3.9.3
django-model-utils==3.1.2
mysqlclient==1.4.2
PyMySQL==0.9.3